+++ lib/fields/credit_card_cvv_field.dart
@@ -45,6 +45,7 @@ class CreditCardCodeField extends StringField {
     super.prefixIcon,
     super.suffix,
     super.suffixIcon,
+    super.onTap,
     super.sizeExtraSmall,
     super.sizeSmall,
     super.sizeMedium,